MM-54201 Move ExplicitMentions to its own file and rename it
